backend/requirements.txt
flask==2.2.5
flask-sqlalchemy==3.0.5
flask-jwt-extended==4.4.4
flask-cors==4.0.0
flask-bcrypt==1.0.1
python-dotenv==0.21.1
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
APScheduler==3.9.1
email-validator==1.3.1
requests==2.28.2
celery==5.2.7
redis==4.3.6
flask-migrate==3.1.0
